requirements.txt
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.1
SQLAlchemy>=1.4
python-dateutil>=2.8.2
pytest>=7.0.0
requests>=2.28.0